Adjusted display contents
- Display task date for the elder dashboard and history page
- Make the status colour circle a bit bigger
- Sorted contact list by message received date
- Make the "sortTasksByDate" a common function and move to common.js
-
