+++ css/dashboard.css
@@ -144,9 +144,14 @@ body {
 .elder .taskCard .statusColor {
   display: inline-block;
   border-radius: 50%;
-  height: 10px;
-  width: 10px;
+  height: 15px;
+  width: 15px;
   margin-right: 0.5rem;
+}
+
+.elder .taskCard p.status {
+  display: flex;
+  align-items: center;
 }
 
 .taskCard .create-button-wrapper {
@@ -342,8 +347,8 @@ button#resetFilter {
 .volunteer .taskCard .statusColor {
   background-color: var(--pureWhite);
   border-radius: 50%;
-  height: 10px;
-  width: 10px;
+  height: 15px;
+  width: 15px;
   position: absolute;
   top: 1.5rem;
   right: 1rem;
