+++ js/history.js
@@ -1,6 +1,6 @@
 import { firestore } from "./firebase/firebase.js";
 import { onSnapshot, collection, query, where } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-firestore.js";
-import { loadPartial, showTabmenu, lazyLoadImages, openModal, closeModal } from "./common.js";
+import { loadPartial, showTabmenu, lazyLoadImages, sortTasksByDate, openModal, closeModal } from "./common.js";
 import { getCurrentUserID, getCurrentUserRole, monitorAuthenticationState } from "./firebase/authentication.js";
 import { getDocument } from "./firebase/firestore.js";
 import { finishLoading, redirect } from "./utils.js";
@@ -241,6 +241,7 @@ function createCardForElder(task) {
   card.innerHTML = `
   <a class="linkURL" href="${task.taskLinkURL}?taskid=${task.taskID}"></a>
   <h3 class="title">${task.taskName}</h3>
+  <p class="date">${task.taskDate}, ${task.taskTime}</p>
   <p class="notes">${task.taskNotes}</p>
   `;
   if (task.taskStatus != "Waiting to be accepted") {
@@ -592,55 +593,5 @@ function setupDateRangePicker() {
 
   function clearDateRangeCalendar(index) {
     setupDateRangePicker();
-    // let today = new Date().toLocaleDateString("en-US", { month: "2-digit", day: "2-digit", year: "numeric" });
-    // let drp = $(".dateRangeModal").eq(index).data("daterangepicker");
-    // drp.setStartDate(today);
-    // drp.setEndDate(today);
-    // drp.updateView();
   }
-}
-
-// TODO: Move this function to a common file
-// ============================================================
-// Utility Functions
-// ============================================================
-
-/**
- * Sorts the given task cards based on the provided date filter value.
- *
- * @param {string} dateFilterValue - The value of the date filter. It can be 'newest' or 'oldest'.
- * @param {NodeList} taskCards - The task cards to be sorted. Each task card is a DOM node.
- *
- * The function first converts the NodeList of task cards into an array. It then sorts the array based on the date attribute of each task card.
- * If the date filter value is 'newest', the task cards are sorted from newest to oldest.
- * If the date filter value is 'oldest', the task cards are sorted from oldest to newest.
- * After sorting, the function replaces the old NodeList in the DOM with the sorted array of task cards.
- */
-function sortTasksByDate(dateFilterValue, taskCards, target) {
-  // Convert NodeList to Array
-  let taskCardsArray = Array.from(taskCards);
-
-  // Sort the array based on the date
-  taskCardsArray.sort((a, b) => {
-    let dateA = new Date(a.getAttribute("data-date"));
-    let dateB = new Date(b.getAttribute("data-date"));
-
-    // For newest to oldest
-    if (dateFilterValue === "newest") {
-      return dateB - dateA;
-    }
-
-    // For oldest to newest
-    if (dateFilterValue === "oldest") {
-      return dateA - dateB;
-    }
-
-    return 0; // If no sorting is needed
-  });
-
-  // Replace the old NodeList with the sorted array
-  // let taskListExplore = document.querySelector("#taskListExplore");
-  taskCardsArray.forEach((card) => {
-    target.appendChild(card);
-  });
 }
